requirements.txt
discord.py==2.0.0
requests
python-dotenv
